requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
Pillow>=10.0.0
gradio>=4.0.0
opencv-python>=4.8.0
einops>=0.7.0
pyyaml>=6.0.1
scikit-image>=0.19.0
tqdm>=4.66.0
matplotlib>=3.5.0
requests>=2.31.0
gdown>=4.7.1
scipy>=1.11.0
psutil>=5.9.0
gputil>=1.4.0 